Refactored ketchup-chart to kup-chart
