+++ packages/ketchup/src/chart.html
@@ -77,19 +77,19 @@
         <hr />
 
         <h1>Horizontal bars</h1>
-        <ketchup-chart id="hbar"></ketchup-chart>
+        <kup-chart id="hbar"></kup-chart>
         <hr />
 
         <h1>Vertical bars</h1>
         <label for="vbar-stacked">Stacked</label>
         <input id="vbar-stacked" type="checkbox" />
-        <ketchup-chart id="vbar"></ketchup-chart>
+        <kup-chart id="vbar"></kup-chart>
         <hr />
 
         <h1>Pie</h1>
         <label for="pie-aspect">3D</label>
         <input id="pie-aspect" type="checkbox" />
-        <ketchup-chart id="pie"></ketchup-chart>
+        <kup-chart id="pie"></kup-chart>
         <hr />
 
         <script src="/assets/chart.js"></script>
