+++ packages/ketchup/src/components/ketchup-chart/readme.md
@@ -1,4 +1,4 @@
-# ketchup-chart
+# kup-chart
